requirements.txt
Flask==2.3.3
pandas==2.1.4
Flask-Session==0.5.0
python-dotenv==1.0.0
requests==2.31.0
Flask-CORS==4.0.0
openpyxl==3.1.2
pymongo==4.6.0
bcrypt==4.1.2
Flask-Mail==0.9.1
itsdangerous==2.1.2
gunicorn
setuptools
wheel

